+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+Release 1.0.8-1
+
+- Improvements to interface staging and diagnostics.
+- Add `--list_interfaces` option to print out interfaces with their respective information
+- For monitor mode, default behavior is to create a vif, except for rtl88XXau cards.
+
 Release 1.0.8
 
 - Interface staging changes (management of a monitor interface) for Intel AX2XX